ai_papers_rag/requirements.txt
streamlit==1.29.0
langchain==0.1.0
langchain-openai==0.0.2
langchain-community==0.0.10
chromadb==0.4.22
pypdf==3.17.4
openai==1.6.1
python-dotenv==1.0.0
tiktoken==0.5.2
numpy==1.24.3
pandas==2.0.3
pydantic==2.5.2
faiss-cpu==1.7.4
sentence-transformers==2.2.2
unstructured==0.11.8
python-multipart==0.0.6
uvicorn==0.24.0
fastapi==0.104.1
rich==13.7.0
tenacity==8.2.3
PyYAML==6.0.1
scikit-learn==1.3.2
plotly==5.17.0